app/requirements.txt
flask
SQLAlchemy
bcrypt
PyJWT